Update: abstract Header component
