+++ src/home/index.js
@@ -6,14 +6,12 @@
  */
 
 import * as React from "react";
-import { Theme, Grid, Icon, Button, Tabs } from "@lugia/lugia-web";
+import { Theme, Grid, Button, Tabs } from "@lugia/lugia-web";
 import Widget from "@lugia/lugia-web/dist/consts/index";
 import styled from "styled-components";
-import { go } from "@lugia/lugiax-router";
-import colorsFunc from "@lugia/lugia-web/dist/css/stateColor";
-import Search from "../search";
+import Header from "../components/header";
+import { linkToUrl } from "../support/commonMethods";
 
-import logo from "../../public/lugia-logo.png";
 import slogan from "../../public/home/slogan.png";
 import logoIntroduction from "../../public/home/logo-introduction.png";
 import designValue from "../../public/home/design-value.png";
@@ -26,7 +24,6 @@ import dynamicLeaf from "../../public/home/dynamic-leaf.gif";
 import banner from "../../public/home/banner.png";
 import backgroundLogo from "../../public/home/background-logo.png";
 
-const { themeColor } = colorsFunc();
 
 const { Row, Col } = Grid;
 
@@ -44,83 +41,6 @@ const BannerImg = styled.div`
   position:absolute;
   top:0;
   z-index:-1;
-`;
-
-const Head = styled.div`
-  height: 80px;
-  line-height: 80px;
-  display: flex;
-  justify-content: space-between;
-`;
-
-const SearchBox = styled.div`
-  height: 100%;
-  padding-top: 10px;
-  float: left;
-`;
-const GuideWrap = styled.div`
-  width: 100%;
-  height: 100%;
-`;
-
-const GitIconContainer = styled.a`
-  height: 80px;
-  width: 90px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  color: #000000;
-  margin-left: 50px;
-  &:hover {
-    color: ${themeColor};
-  }
-`;
-const GitIcon = styled.div`
-  width: 28px;
-  height: 28px;
-  font-size: 16px;
-  border-radius: 50%;
-  box-shadow: 0 0 6px 0 rgba(51, 51, 51, 0.3);
-  display: grid;
-  place-items: center;
-`;
-const GitStar = styled.div`
-  width: 55px;
-  padding: 0 10px;
-  margin-left: 8px;
-  height: 30px;
-  font-size: 14px;
-  font-weight: bold;
-  border-radius: 4px;
-  background: #f2f2f2;
-  text-align: center;
-  line-height: 30px;
-  position: relative;
-  &:before {
-    content: "";
-    display: block;
-    position: absolute;
-    z-index: -1;
-    left: -12px;
-    top: 7px;
-    width: 0px;
-    height: 0px;
-    border-top: 8px solid transparent;
-    border-right: 8px solid #f2f2f2;
-    border-bottom: 8px solid transparent;
-    border-left: 8px solid transparent;
-  }
-`;
-
-const HeadIndex = styled.a`
-  height: 100%;
-  float: right;
-  margin-left: 50px;
-  font-size: 14px;
-  color: #50575d;
-  &:hover {
-    color: ${themeColor};
-  }
 `;
 
 const SloganWrap = styled.div`
@@ -317,13 +237,6 @@ const SolutionItemContent = styled.div`
   justify-content: space-between;
 `;
 
-const Logo = styled.img`
-  height: 34px;
-  margin: 24px 0 0 0;
-  float: left;
-  cursor: pointer;
-`;
-
 const TabsBlock = styled.div`
   font-family: PingFangSC-Regular;
   width: 100%;
@@ -383,10 +296,6 @@ const CopyRight = styled.div`
   color: #747e90;
   text-align: center;
 `;
-
-function linkToUrl(target: string) {
-  target && go({ url: target });
-}
 
 const solutionData = [
   {
@@ -531,11 +440,9 @@ export default class Pages extends React.Component<any, any> {
 
   constructor(props) {
     super(props);
-    this.asyncGetStar();
   }
 
   render() {
-    const { stars } = this.state;
     return (
       <React.Fragment>
         <BannerImg />
@@ -544,46 +451,7 @@ export default class Pages extends React.Component<any, any> {
             <Wrapper> </Wrapper>
           </Col>
           <Col span={22} md={{ span: 22 }} xl={{ span: 20 }} xxl={{ span: 18 }}>
-            <Head title="head">
-              <Logo src={logo} alt="" onClick={() => linkToUrl("/home")} />
-              <SearchBox>
-                <Search />
-              </SearchBox>
-              <GuideWrap>
-                <Row>
-                  <Col
-                    span={0}
-                    lg={{ span: 0 }}
-                    xl={{ span: 24 }}
-                    xxl={{ span: 24 }}
-                  >
-                    <HeadIndex onClick={() => linkToUrl("/logs")}>
-                      更新日志
-                    </HeadIndex>
-                    <HeadIndex onClick={() => linkToUrl("/lugia-mega")}>
-                      lugia-mega
-                    </HeadIndex>
-                    <HeadIndex onClick={() => linkToUrl("/component")}>
-                      lugia-web
-                    </HeadIndex>
-                    <HeadIndex onClick={() => linkToUrl("/component/affix")}>
-                      设计指南
-                    </HeadIndex>
-                  </Col>
-                </Row>
-              </GuideWrap>
-
-              <GitIconContainer
-                href={"https://github.com/lugia-ysstech/lugia"}
-                target={"_blank"}
-              >
-                <GitIcon>
-                  {" "}
-                  <Icon iconClass={"lugia-icon-logo_github"} />{" "}
-                </GitIcon>
-                <GitStar>{stars}</GitStar>
-              </GitIconContainer>
-            </Head>
+            <Header />
             <SloganWrap>
               <Row>
                 <Col
@@ -755,23 +623,5 @@ export default class Pages extends React.Component<any, any> {
         </Footer>
       </React.Fragment>
     );
-  }
-
-  async asyncGetStar() {
-    const result = await fetch(
-      "https://api.github.com/repos/lugia-ysstech/lugia",
-      {
-        method: "GET",
-        headers: new Headers({ "Content-Type": "application/json" })
-      }
-    )
-      .then(response => response.json())
-      .then(data => {
-        return data;
-      });
-    const { stargazers_count } = result;
-    this.setState({
-      stars: stargazers_count
-    });
   }
 }
