+++ src/support/commonMethods.js
@@ -1,0 +1,5 @@
+import { go } from "@lugia/lugiax-router";
+
+export const linkToUrl = (target: string) => {
+  target && go({ url: target });
+};
